[Feat] 회원가입 스크린을 구현하고,  최초 진입 페이지를 수정한다.
